backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
openai==1.58.1
yfinance==0.2.51
pandas==2.2.3
numpy==2.2.1
Pillow==11.1.0
tavily-python==0.5.0
fpdf2==2.8.2
pydantic==2.10.4
python-multipart==0.0.20
sse-starlette==2.2.1
python-dotenv==1.0.1
